% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_interpolate.R
\name{dbi_interpolate}
\alias{dbi_interpolate}
\title{Wrapper for DBIs sqlInterpolate}
\usage{
dbi_interpolate(value, quote = NULL)
}
\arguments{
\item{value}{The value to be quoted.}

\item{quote}{Optional quoting character to use; can be one of '[' and '"'.
If this is NULL, DBI's default is used.}
}
\description{
This function uses `sqlInterpolate` to parse character values.
The quote par
}
