% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Get Parameter Names}
\usage{
get_params(.query)
}
\arguments{
\item{.query}{SQL text}
}
\description{
Retrieve parameter names from SQL text. Assumes all parameters are of the
form "@param".
}
