% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_set.R
\name{IN}
\alias{IN}
\title{Mark parameter value for use with IN clause}
\usage{
IN(x = NULL)
}
\arguments{
\item{x}{a vector of values, possibly only of length 1}
}
\value{
The original object with the class attribute appended with "IN".
}
\description{
Mark parameter value for use with IN clause
}
