% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_set.R
\name{IS_NULL}
\alias{IS_NULL}
\title{Flag parameter to be set to NULL}
\usage{
IS_NULL()
}
\description{
Called with no arguments, this flags a parameter for conversion to NULL,
i.e. from something like \code{column <= @param} to \code{column is null}.
}
