% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_docs.R
\name{is_null_string}
\alias{is_null_string}
\title{Does the string contain no matter, but very well [:space:]?}
\usage{
is_null_string(string)
}
\arguments{
\item{string}{the string to check}
}
\value{
TRUE if the string contains words, otherwise FALSE
}
\description{
Does the string contain no matter, but very well [:space:]?
}
