% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_scalar_character.R
\name{is_scalar_character}
\alias{is_scalar_character}
\title{Is a Value a Scalar of Character Type?}
\usage{
is_scalar_character(value)
}
\arguments{
\item{value}{A value to be tested.}
}
\value{
logical
}
\description{
Is a Value a Scalar of Character Type?
}
