% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_set.R
\name{LIKE}
\alias{LIKE}
\title{Wrap parameter value in wildcards}
\usage{
LIKE(x, side, wildcard = "\%")
}
\arguments{
\item{x}{character}

\item{side}{character; one of 'l' (left), 'r' (right) or 'b' (both)}

\item{wildcard}{character; defaults to '%'}
}
\description{
Pad value with wildcard symbol for when it is used with a parameter in a
LIKE expression in SQL.
}
