% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_packaged.R
\name{package_name}
\alias{package_name}
\title{Get Package Name}
\usage{
package_name()
}
\value{
logical
}
\description{
Find the name of the first non-squr package.
If not packaged, an empty character value is returned.
}
