% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_docs.R
\name{parse_instruduction}
\alias{parse_instruduction}
\title{Parse introduction: the premier part of a roxygen block
containing description and option details separated by
a blank roxygen line.}
\usage{
parse_instruduction(expression)
}
\arguments{
\item{expression}{the description to be parsed}
}
\value{
A list containing the parsed description
}
\description{
Copied from roxygen2: https://github.com/yihui/roxygen2/blob/master/R/parse-preref.R
}
\author{
yihui
}
