% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{+.sq}
\alias{+.sq}
\title{Concatenate sq objects}
\usage{
\method{+}{sq}(e1, e2)
}
\arguments{
\item{e1}{An \code{sq} object.}

\item{e2}{An \code{sq} object.}
}
\value{
An \code{sq} object.
}
\description{
Concatenate two sq chunks, separated by two line breaks.
}

