% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_docs.R
\name{print.sq_doc}
\alias{print.sq_doc}
\title{Prints the SQL documentation.}
\usage{
\method{print}{sq_doc}(x, ...)
}
\arguments{
\item{x}{sq_doc object.}

\item{...}{currently unused.}
}
\description{
Prints the SQL documentation.
}
