% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sq}
\alias{print.sq}
\title{S3 Print Method for 'sq' Objects}
\usage{
\method{print}{sq}(x, ...)
}
\arguments{
\item{x}{an sq object to be printed.}

\item{...}{Passed on to \code{cat}.}
}
\description{
S3 Print Method for 'sq' Objects
}

