% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sq_docs}
\alias{print.sq_docs}
\title{S3 Print method for 'sq_docs' objects}
\usage{
\method{print}{sq_docs}(docs, ...)
}
\arguments{
\item{docs}{a sq_docs object generated by sq_parse_docs}
}
\description{
S3 Print method for 'sq_docs' objects
}
