% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sq_value}
\alias{print.sq_value}
\title{S3 Print Method for 'sq_value' Objects}
\usage{
\method{print}{sq_value}(x, ...)
}
\arguments{
\item{x}{an sq_value object to be printed}

\item{...}{Passed to \code{print.default}.}
}
\description{
S3 Print Method for 'sq_value' Objects
}

