% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql_file.R
\name{read_sql_file}
\alias{read_sql_file}
\title{Read a SQL File}
\usage{
read_sql_file(path)
}
\arguments{
\item{path}{The path to the file to be read.}
}
\value{
List with elements \code{sql} containing the SQL text and \code{docs}
containing any documention lines beginning with #'.
}
\description{
Read a SQL File
}
