% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_bind.R
\name{sq_bind}
\alias{sq_bind}
\title{Bind values to parameters in SQL}
\usage{
sq_bind(.query)
}
\arguments{
\item{.query}{A sq object}
}
\value{
A list, possibly of length greater than one, of SQL queries
}
\description{
Insert values into SQL parameters.
}
