% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_bind.R
\name{sq_bind_}
\alias{sq_bind_}
\title{Internal Recursive Binding Function}
\usage{
sq_bind_(query, params)
}
\arguments{
\item{query}{The query text to be parameterized.}

\item{params}{A list of parameters.}
}
\description{
Internal Recursive Binding Function
}
\details{
The parameters should be sorted to have longer names first
  to avoid errors when some names are subsets of others.
}
