% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_docs.R
\name{sq_docs}
\alias{sq_docs}
\title{Parses the query and returns a list with all the elements of the comment.}
\usage{
sq_docs(.query)
}
\arguments{
\item{query_docs}{the query name.}
}
\value{
a list with documentation including \code{introduction}, \code{return},
        and \code{params} (as a data frame).
}
\description{
Parses the query and returns a list with all the elements of the comment.
}
