% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_docs.R
\name{sq_parse_docs}
\alias{sq_parse_docs}
\title{Parse SQL file documentation}
\usage{
sq_parse_docs(.query)
}
\arguments{
\item{.query}{a \code{sq} object, e.g. created by \code{sq_file}}
}
\value{
list with class "sq_docs" containing the documentation parsed
into markdown
}
\description{
Parse the documentation (if any exists) from the header of the SQL file
that generated \code{.query}.
}
\details{
See \link{\code{sq_file}} for a description of the documentation
format. When a "sq_docs" is printed at the console or via \code{sq_view_docs}
the markdown is written to a temp file, converted to html and then viewed
in either Rstudio's viewer pane (if available) or via \code{browseURL}.
}
