% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_replace.R
\name{sq_replace}
\alias{sq_replace}
\title{Replace Arbitary Parts of Query Using gsub}
\usage{
sq_replace(.query, pattern, replacement, ...)
}
\arguments{
\item{pattern}{character; passed to pattern argument of gsub}

\item{replacement}{character; passed to replacement argument of gsub}

\item{\dots}{further arguments to gsub}

\item{sq}{squr query object}
}
\description{
Replace Arbitary Parts of Query Using gsub
}
