% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_send.R
\name{sq_send}
\alias{sq_send}
\title{Send a Query}
\usage{
sq_send(.query, .with, ...)
}
\arguments{
\item{.query}{A character/sq query string.}

\item{.with}{function with which to send the query.}

\item{...}{further parameters passed to \code{with}.}
}
\value{
The value returned by \code{.with}
}
\description{
Small function that invokes another given function for sending
a query to the database. The benefit of using this wrapper is
to keep database/driver boilerplate code separate from functions
that sends queries.
}
