% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_set.R
\name{sq_set_}
\alias{sq_set_}
\title{Internal Recursive Parameterization Function}
\usage{
sq_set_(query, params)
}
\arguments{
\item{query}{The query text to be parameterized.}

\item{params}{A list of parameters.}
}
\description{
Internal Recursive Parameterization Function
}
\details{
The parameters should be sorted to have longer names first
  to avoid errors when some names are subsets of others.
}
