% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_set_include.R
\name{sq_set_include}
\alias{sq_set_include}
\title{Parameterize Insert Parameters}
\usage{
sq_set_include(query, label, path)
}
\arguments{
\item{query}{An \code{sq} query object.}

\item{label}{the insert block label.}

\item{path}{see \code{sq_file}. As alternative to a path, one can
also supply an \code{sq} object directly, which is used as is.}
}
\description{
Parameterize Insert Parameters
}
\details{
The values will be prepared with \code{sq_value}
}

