% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_text.R
\name{sq_text}
\alias{sq_text}
\title{Read The Contents of an SQL File}
\usage{
sq_text(text)
}
\arguments{
\item{text}{A character SQL query.}
}
\value{
An \code{sq} object.
}
\description{
Read The Contents of an SQL File
}

