% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_transaction.R
\name{sq_transaction}
\alias{sq_transaction}
\title{Wrap SQL Queries in a Transaction}
\usage{
sq_transaction(..., .name = "")
}
\arguments{
\item{...}{\code{sq} objects.}

\item{.name}{Optional name for the transaction.}
}
\description{
The provided queries are concatenated and wrapped in
\code{BEGIN TRANSACTION ... COMMIT TRANSACTION}.
}

