% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_set.R
\name{LIKE}
\alias{LIKE}
\title{Pad values with wildcard symbol}
\usage{
LIKE(x = NULL, side, wildcard = "\%")
}
\arguments{
\item{x}{character}

\item{side}{character; one of 'l' (left), 'r' (right) or 'b' (both)}

\item{wildcard}{character; defaults to "\%"}
}
\description{
Pads single values with wildcard symbol on left, right or both
for use with parameters in LIKE expressions.
}
