% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_file.R
\name{sq_file}
\alias{sq_file}
\title{Read the contents of an SQL file}
\usage{
sq_file(path, override_pkg = FALSE)
}
\arguments{
\item{path}{character specifying the path to an SQL file. The ".sql"
extension can be omitted if and only if the actual extension is lower case.
When used in a package, the path will be taken to be relative to the
\code{inst} folder.}

\item{override_pkg}{boolean if \code{TRUE} then forces the use of the raw path
rather than looking inside the package installation.}
}
\value{
A \code{sq} object; a list with components \code{sql}, \code{params},
\code{values} and \code{docs}.
}
\description{
Read the contents of an SQL file
}
\details{
\code{sq_file} assumes that any lines at the beginning of the SQL file
that are commented out via "- -" (two dashes) contain documentation. Subsequent commented out
lines within the query itself are not considered part of the SQL file documentation.

It assumes a simple format roughly in the style of roxygen2. The first line
contains the query title, followed by a blank comment line, and then possibly
several lines comprising a description of the query.

Next, \code{sq_file} recognizes three tags:

\describe{
\item{@param}{the name of the parameter followed by a space and then a description;
you can include multiple param tags each starting on its own line}
\item{@functions}{comma separated list of functions that use this query}
\item{@scripts}{comma separated list of scripts that use this query}
}
}
