% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_file.R
\name{sq_file}
\alias{sq_file}
\title{Read The Contents of an SQL File}
\usage{
sq_file(path)
}
\arguments{
\item{path}{character specifying the path to an SQL file. The ".sql" extension can be omitted iff
the actual extension is lower case.
When used in a package, the path will be taken to be relative to the \code{inst} folder.}
}
\value{
A \code{sql} object.
}
\description{
Read The Contents of an SQL File
}

