% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_set.R
\name{sq_set}
\alias{sq_set}
\title{Parameterize an SQL Query}
\usage{
sq_set(.query, ...)
}
\arguments{
\item{.query}{An \code{sq} object or a character string.}

\item{...}{name-value pairs of named parameters.}
}
\description{
Parameterize an SQL Query
}
\details{
The values will be prepared with \code{sq_value}
}
