% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_set_insert.R
\name{sq_set_insert}
\alias{sq_set_insert}
\title{Parameterize Insert Parameters}
\usage{
sq_set_insert(.query, .label, .into, ..., .data = parent.frame(),
  .split = 75, .quote = "[")
}
\arguments{
\item{.query}{An \code{sq} query object.}

\item{.label}{the insert block label.}

\item{.into}{The name of the table to insert into.}

\item{...}{Names of the values to insert, either as character or one-sided formulas.
It is also possible to specify \code{AsIs} variables by wrapping the value in \code{I(.)}.}

\item{.data}{Data or environment to lookup the names.}

\item{.split}{integer specifying the maximum number of value-pairs in each \code{INSERT} block.}

\item{.quote}{the quote to use for table and column names.}
}
\description{
Parameterize Insert Parameters
}
\details{
The values will be prepared with \code{sq_value}
}

