% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_value.R
\name{sq_value}
\alias{sq_value}
\title{Prepare Value for an SQL Query}
\usage{
sq_value(value, quote = "'")
}
\arguments{
\item{value}{A value to be used in an SQL query.}

\item{quote}{The character to be used to quote the (non-numeric) value(s).
For brackets, use the opening bracket.}
}
\value{
character: A character representation appropriate for SQL queries.
}
\description{
Prepare Value for an SQL Query
}
\details{
\code{NA}s are converted to NULL, other values are coerced to character representation
 with \code{as.character}, and non-numeric and non-NULL values are (single) quoted. Lists
 are wrapped in parentheses and values are separated with commas. The \code{quote} parameter
 is useful for dynamically specifying e.g. columns, or table names.
}

